backend/requirements.txt
# Web框架
fastapi==0.109.0
uvicorn==0.27.0

# 调试工具
debugpy>=1.6.7
python-dotenv>=1.0.0

# 测试工具
pytest==7.4.4
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-cov>=3.0.0

# 代码质量工具
black>=22.3.0
pylint>=2.15.0
isort>=5.10.0

# 数据库
sqlalchemy>=2.0.0
alembic>=1.11.0
pymysql>=1.0.2

# 工具包
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiofiles>=0.8.0

# 自动化测试
selenium>=4.1.0
playwright==1.41.2
appium-python-client>=2.6.0
pytest-playwright>=0.7.0
pytest-selenium>=4.0.0
pytest-xdist>=2.5.0

# JWT相关依赖
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# 缓存和安全相关
redis>=4.6.0
aioredis>=2.0.0
captcha>=0.4
prometheus-client>=0.17.0
starlette-prometheus>=0.9.0

# Appium相关依赖
Appium-Python-Client==3.1.1
selenium==4.16.0
PyYAML==6.0.1 